Fix uint8 overflowing in uart

+++ src/libsc/k60/uart_device.cpp
@@ -139,7 +139,7 @@ public:
 			kVector,
 		} type;
 		size_t size;
-		volatile uint8_t it;
+		volatile size_t it;
 		bool is_mem_owned;
 	};
